test: split supervised exec tests by behavior
