+++ crates/vsock-host/src/tests/exec_operation/supervised/mod.rs
@@ -1,0 +1,7 @@
+mod cancel;
+mod control;
+mod lifecycle;
+mod output;
+mod support;
+mod timeout;
+mod validation;
